requirements.txt
absl-py==2.1.0
alembic==1.13.1
beautifulsoup4==4.12.3
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
contourpy==1.1.1
cycler==0.12.1
databricks-cli==0.18.0
docker==6.1.3
entrypoints==0.4
filelock==3.13.1
flask==3.0.1
fonttools==4.47.2
gdown==5.0.1
gitdb==4.0.11
GitPython==3.1.41
google-auth==2.27.0
google-auth-oauthlib==1.0.0
greenlet==3.0.3
grpcio==1.60.0
gunicorn==21.2.0
idna==3.6
importlib-metadata==7.0.1
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
kiwisolver==1.4.5
Mako==1.3.1
Markdown==3.5.2
MarkupSafe==2.1.4
matplotlib==3.7.4
mlflow==2.9.2
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
oauthlib==3.2.2
opencv-python==4.9.0.80
openslide-python==1.3.1
packaging==23.2
pandas==2.0.3
pillow==10.2.0
poptorch==3.3.0+113432
protobuf==4.25.2
pyarrow==14.0.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyJWT==2.8.0
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
querystring-parser==1.2.4
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
smmap==5.0.1
soupsieve==2.5
SQLAlchemy==2.0.25
sqlparse==0.4.4
sympy==1.12
tabulate==0.9.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
threadpoolctl==3.2.0
torch==2.0.1+cpu
torchaudio==2.0.2+cpu
torchvision==0.15.2+cpu
tqdm==4.66.1
typing-extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
websocket-client==1.7.0
werkzeug==3.0.1
zipp==3.17.0
